+++ src/api/consumer/lib/exceptions/id-exported-already.js
@@ -1,0 +1,10 @@
+/** @flow */
+export default class IdExportedAlready extends Error {
+  id: string;
+
+  constructor(id: string, remote: string) {
+    super();
+    this.id = id;
+    this.remote = remote;
+  }
+}