+++ src/cli/default-error-handler.js
@@ -26,6 +26,7 @@ import RemoteNotFound from '../remotes/exceptions/remote-not-found';
 import { ScopeNotFound, ResolutionException, ComponentNotFound, DependencyNotFound } from '../scope/exceptions';
 import InvalidBitJson from '../consumer/bit-json/exceptions/invalid-bit-json';
 import invalidIdOnCommit from '../api/consumer/lib/exceptions/invalid-id-on-commit';
+import IdExportedAlready from '../api/consumer/lib/exceptions/id-exported-already';
 import PathNotExists from '../api/consumer/lib/exceptions/path-not-exists';
 import FileSourceNotFound from '../consumer/component/exceptions/file-source-not-found';
 import { MissingMainFile, MissingBitMapComponent } from '../consumer/bit-map/exceptions';
@@ -68,6 +69,7 @@ const errorsMap: [[Error, (err: Error) => string]] = [
   [ MissingMainFile, err => `fatal: the main file ${chalk.bold(err.mainFile)} was not found in the files list ${chalk.bold(err.files.join(', '))}`],
   [ MissingBitMapComponent, err => `fatal: the component ${chalk.bold(err.id)} was not found in the bit.map file`],
   [ PathNotExists, err => `fatal: the file "${chalk.bold(err.path)}" was not found`],
+  [ IdExportedAlready, err => `error - the component ${chalk.bold(err.id)} has been already exported to ${chalk.bold(err.remote)}`],
   [ invalidIdOnCommit, err => `error - Unable to commit. ${chalk.bold(err.id)} not found.
 Run \`bit status\` command to list all components available for commit.`]
 ];